requirements.txt
toml
pynput
# lyricsgenius
dbus-python